Fix wrong Sponza location as in #148
As of now, when building the entirety of the OWL repo (including the
samples) under Windows with VS, the executable is not generated under
the same directory of the sample, but all executables end up in the
same directory (e.g. build/Debug).

As a result, the hardcoded path that the samples/advanced/optix7course
sample uses to find the Sponza model does not work. This commit
makes CMake set up a variable that signals whether the whole project
has been built, and in that case the sample looks for the appropriate
reative path for its location.