Remove include path /usr/local/cuda/include/.
On a system where several nvcc toolkit are installed, this path is often an
alias to the latest installed toolkit; hence when trying to build with an older
version off nvcc you end up in a situation where the old nvcc compiler is using
new header; this situation may lead to error at link time (undefined symbols).
See issue #191 for discussion.